% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howBadIsIt.R
\name{howBadIsIt}
\alias{howBadIsIt}
\title{\code{howBadIsIt}}
\usage{
howBadIsIt(dat)
}
\arguments{
\item{dat}{string. A data.frame containing two columns: R, which is denotes sample membership and must take values {0,1};
and Y, which is the variable of interest. The function will throw an error is colnames(dat) != c("R","Y").}
}
\value{
A data.frame with four columns: d1, which is the empirical correlation between sample membership and the variable of interest; estimated_d1,
which is an estimate of d1 obtained by solving Meng's formula for estimation error for d1 (this value reflects selection bias plus other sources of error in the data);
d2, which is sqrt((1-f)/f), i.e. the data quantity term; d3, which is the problem difficulty term, sd(Y); error, which
is the difference between the sample average and population average, given by d1*d2*d3; and finally, nEff, which is the effective size of the sample, that is, the
size of the random sample that would yield an estimate of the population average with same error.
}
\description{
Given a dataset, this function calculates the correlation between sample membership and the variable of interest, the effective sample size, and more.
}
\details{
The function returns various quantities of interest defined in Meng, X. L. (2018). Statistical paradises and paradoxes in big data (I): Law of large populations, big data paradox, and the 2016 us presidential election. Annals of Applied Statistics, 12(2), 685-726. https://doi.org/10.1214/18-AOAS1161SF
}
