% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSamples.R
\name{simulateSamples}
\alias{simulateSamples}
\title{\code{simulateSamples}}
\usage{
simulateSamples(dat, N, n, n_eff, ddc, nSim)
}
\arguments{
\item{dat}{string. A data.frame containing two columns: R, which is denotes sample membership and must take values {0,1};
and Y, which is the variable of interest. The function will throw an error is colnames(dat) != c("R","Y").}

\item{N}{Numeric. Population size.}

\item{n}{Numeric. Sample size.}

\item{n_eff}{Numeric. Effective sample size. Might be taken from the outputs of \code{howBadIsIt}.}

\item{ddc}{Numeric. Data defect correlation. Might be taken from the outputs of \code{howBadIsIt}.}

\item{nSim}{Numeric. Number of samples to simulate. nSim SRSs and large, biased samples will be simulated.}
}
\value{
A data.frame with eight columns: smallSRSEstimate = Estimate of the population mean from the SRS of size n_eff; largeBiasedEstimate = estimate of population mean from the biased sample of size n;
smallSRSCoverage = % of (Wald) confidence intervals from the simulated SRSs covering the true population mean;
largeBiasedCoverage =  % of (Wald) confidence intervals from the simulated biased samples of size n covering the true population mean;
smallSRSError = sample mean - population mean for the SRS samples; largeBiasedError = sample mean - population mean for the biased samples of size n;
smallSRSSquaredError = smallSRSError squared; largeBiasedSquaredError = largeBiasedError squared; sim = simulation number. The data.frame has nSim rows, one for each simulation.
}
\description{
Simulates many simple random samples (SRSs) of size n_eff and larger, based samples of size n. n_eff, n and other arguments may be taken from the outputs of \code{howBadIsIt}. For each sample,
the function returns various statistics (see below).
}
\details{
The function returns various quantities of interest defined in Meng, X. L. (2018). Statistical paradises and paradoxes in big data (I): Law of large populations, big data paradox, and the 2016 us presidential election. Annals of Applied Statistics, 12(2), 685-726. https://doi.org/10.1214/18-AOAS1161SF
}
